move share entity, DO, repo and tests to share module
