+++ apps/server/src/modules/sharing/controller/share-token.controller.spec.ts
@@ -1,8 +1,9 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { MikroORM } from '@mikro-orm/core';
 import { Test, TestingModule } from '@nestjs/testing';
-import { CopyElementType, CopyStatus, CopyStatusEnum, ICurrentUser, ShareTokenParentType } from '@shared/domain';
+import { CopyElementType, CopyStatus, CopyStatusEnum, ICurrentUser } from '@shared/domain';
 import { courseFactory, setupEntities, shareTokenFactory } from '@shared/testing';
+import { ShareTokenParentType } from '../domainobject/share-token.do';
 import { ShareTokenUC } from '../uc';
 import { ShareTokenInfoDto } from '../uc/dto';
 import { ShareTokenController } from './share-token.controller';
