+++ apps/server/src/modules/sharing/service/token-generator.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { ShareTokenString } from '@shared/domain';
 import { nanoid } from 'nanoid';
+import { ShareTokenString } from '../domainobject/share-token.do';
 
 @Injectable()
 export class TokenGenerator {
