+++ apps/server/src/modules/sharing/sharing.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
-import { CourseRepo, ShareTokenRepo } from '@shared/repo';
+import { CourseRepo } from '@shared/repo';
 import { LoggerModule } from '@src/core/logger';
 import { AuthorizationModule } from '@src/modules/authorization';
 import { LearnroomModule } from '@src/modules/learnroom';
 import { ShareTokenController } from './controller/share-token.controller';
-import { ShareTokenService, TokenGenerator } from './service';
 import { ShareTokenUC } from './uc';
+import { ShareTokenService, TokenGenerator } from './service';
+import { ShareTokenRepo } from './repo/share-token.repo';
 
 @Module({
 	imports: [AuthorizationModule, LearnroomModule, LoggerModule],
