+++ apps/server/src/shared/domain/domainobject/index.ts
@@ -1,4 +1,3 @@
 export * from './base.do';
 export * from './pseudonym.do';
-export * from './share-token.do';
 export * from './video-conference.do';
