+++ apps/server/src/shared/testing/factory/share-token.do.factory.ts
@@ -1,5 +1,6 @@
 /* istanbul ignore file */
-import { EntityId, ShareTokenDO, ShareTokenParentType } from '@shared/domain';
+import { EntityId } from '@shared/domain';
+import { ShareTokenDO, ShareTokenParentType } from '@src/modules/sharing/domainobject/share-token.do';
 import { ObjectId } from 'bson';
 import { Factory } from 'fishery';
